Fixes.. Change DAO pattern and use JDBI object directly instead of 'onDemand'
There is an unimplemented method which is causing errors.
This update fixes that by converting a method that needed annotations as
an interface to instead a void method of a POJO (which does not
cause any errors). For follow up, we should implement observability, ie:
a way for users to see the results of map indexing (eg: a webpage)